requirements.txt
flask
chromadb
anthropic
python-dotenv